feat: use VirusTotal v3 API
Use VirusTotal v3 API completely instead of querying both the v2
and the v3 API.
Also: Separate Munin into multiple files to make it clearer which
function is used where.
