+++ munin.py
@@ -28,9 +28,12 @@ import argparse
 import logging
 import gzip
 from datetime import datetime
-from future.utils import viewitems
 from colorama import init, Fore, Back, Style
 from lib.helper import generateResultFilename
+import lib.munin_vt as munin_vt
+from lib.munin_csv import writeCSVHeader, writeCSV, CSV_FIELDS
+import lib.connections as connections
+from lib.munin_stdout import printResult, printHighlighted, printKeyLine
 import cfscrape
 # Handle modules that may be difficult to install
 # e.g. pymisp has no Debian package, selenium is obsolete
@@ -55,64 +58,13 @@ except ImportError as e:
 # CONFIG ##############################################################
 
 # API keys / secrets - please configure them in 'munin.ini'
-VT_PUBLIC_API_KEY = '-'
 MAL_SHARE_API_KEY = '-'
 PAYLOAD_SEC_API_KEY = '-'
-PROXY = '-'
-
-VENDORS = ['Microsoft', 'Kaspersky', 'McAfee', 'CrowdStrike', 'TrendMicro',
-           'ESET-NOD32', 'Symantec', 'F-Secure', 'Sophos', 'GData']
 
 WAIT_TIME = 16  # Public API allows 4 request per minute, so we wait 15 secs by default
 
-CSV_FIELD_ORDER = ['Lookup Hash', 'Rating', 'Comment', 'Positives', 'Virus', 'File Names', 'First Submitted',
-                   'Last Submitted', 'File Type', 'MD5', 'SHA1', 'SHA256', 'Imphash', 'Harmless', 'Revoked',
-                   'Expired', 'Trusted', 'Signed', 'Signer', 'Hybrid Analysis Sample', 'MalShare Sample',
-                   'VirusBay Sample', 'MISP', 'MISP Events', 'URLhaus', 'AnyRun', 'CAPE', 'VALHALLA', 'User Comments']
-
-CSV_FIELDS = {'Lookup Hash': 'hash',
-              'Rating': 'rating',
-              'Comment': 'comment',
-              'Positives': 'positives',
-              'Virus': 'virus',
-              'File Names': 'filenames',
-              'First Submitted': 'first_submitted',
-              'Last Submitted': 'last_submitted',
-              'File Type': 'filetype',
-              'File Size': 'filesize',
-              'MD5': 'md5',
-              'SHA1': 'sha1',
-              'SHA256': 'sha256',
-              'Imphash': 'imphash',
-              'Harmless': 'harmless',
-              'Revoked': 'revoked',
-              'Expired': 'expired',
-              'Trusted': 'mssoft',
-              'Signed': 'signed',
-              'Signer': 'signer',
-              'Hybrid Analysis Sample': 'hybrid_available',
-              'MalShare Sample': 'malshare_available',
-              'VirusBay Sample': 'virusbay_available',
-              'MISP': 'misp_available',
-              'MISP Events': 'misp_events',
-              'URLhaus': 'urlhaus_available',
-              'AnyRun': 'anyrun_available',
-              'CAPE': 'cape_available',
-              'VALHALLA': 'valhalla_matches',
-              'Comments': 'comments',
-              'User Comments': 'commenter',
-              'Reputation': 'reputation',
-              'Times Submitted': 'times_submitted',
-              'Tags': 'tags',
-              }
-
 TAGS = ['HARMLESS', 'SIGNED', 'MSSOFT', 'REVOKED', 'EXPIRED']
 
-# VirusTotal URLs
-VT_REPORT_URL = 'https://www.virustotal.com/vtapi/v2/file/report'
-VT_COMMENT_API = 'https://www.virustotal.com/ui/files/%s/comments?relationships=item,author'
-VT_PERMALINK_URL = 'https://www.virustotal.com/gui/file/%s/detection'
-VT_DETAILS = 'https://www.virustotal.com/ui/files/%s'
 # MalwareShare URL
 MAL_SHARE_URL = 'http://malshare.com/api.php'
 # Hybrid Analysis URL
@@ -190,7 +142,7 @@ def processLine(line, debug):
 
         # Get Information
         # Virustotal
-        vt_info = getVTInfo(hashVal)
+        vt_info = munin_vt.getVTInfo(hashVal)
         info.update(vt_info)
         # MISP
         misp_info = getMISPInfo(hashVal)
@@ -270,7 +222,7 @@ def processLines(lines, resultFile, nocsv=False, debug=False):
 
         # Comment on Sample
         if args.comment and info['sha256'] != "-":
-            commentVTSample(info['sha256'], "%s %s" % (args.p, info['comment']))
+            munin_vt.commentVTSample(info['sha256'], "%s %s" % (args.p, info['comment']))
 
         # Download Samples
         if args.download and 'sha256' in info:
@@ -313,213 +265,6 @@ def fetchHash(line):
         return hash, hashTypes[len(hash)], rest
     return '', '', ''
 
-
-def getVTInfo(hash):
-    """
-    Retrieves many different attributes of a sample from Virustotal via its hash
-    :param hash:
-    :return:
-    """
-    # Sample info
-    sample_info = {
-        'hash': hash,
-        "result": "- / -",
-        "virus": "-",
-        "last_submitted": "-",
-        "first_submitted": "-",
-        "filenames": "-",
-        "filetype": "-",
-        "rating": "unknown",
-        "positives": 0,
-        "res_color": Back.CYAN,
-        "imphash": "-",
-        "harmless": False,
-        "revoked": False,
-        "signed": False,
-        "expired": False,
-        "mssoft": False,
-        "vendor_results": {},
-        "signer": "",
-        "comments": 0,
-        "commenter": [],
-        "vt_queried": True,
-        "tags": [],
-    }
-
-    # Prepare VT API request
-    parameters = {"resource": hash, "apikey": VT_PUBLIC_API_KEY}
-    success = False
-    while not success:
-        try:
-            response_dict_code = requests.get(VT_REPORT_URL, params=parameters, proxies=PROXY)
-            response_dict = json.loads(response_dict_code.content.decode("utf-8"))
-            success = True
-        except Exception as e:
-            if args.debug:
-                traceback.print_exc()
-                # print "Error requesting VT results"
-            pass
-
-    sample_info['vt_verbose_msg'] = response_dict.get("verbose_msg")
-
-    if response_dict.get("response_code") > 0:
-        # Hashes
-        sample_info["md5"] = response_dict.get("md5")
-        sample_info["sha1"] = response_dict.get("sha1")
-        sample_info["sha256"] = response_dict.get("sha256")
-        # AV matches
-        sample_info["positives"] = response_dict.get("positives")
-        sample_info["total"] = response_dict.get("total")
-        sample_info["last_submitted"] = response_dict.get("scan_date")
-        # Virus Name
-        scans = response_dict.get("scans")
-        virus_names = []
-        sample_info["vendor_results"] = {}
-        for vendor in VENDORS:
-            if vendor in scans:
-                if scans[vendor]["result"]:
-                    virus_names.append("{0}: {1}".format(vendor, scans[vendor]["result"]))
-                    sample_info["vendor_results"][vendor] = scans[vendor]["result"]
-                else:
-                    sample_info["vendor_results"][vendor] = "-"
-            else:
-                sample_info["vendor_results"][vendor] = "-"
-
-        if len(virus_names) > 0:
-            sample_info["virus"] = " / ".join(virus_names)
-
-        # Positives / Total
-        sample_info['vt_positives'] = response_dict.get("positives")
-        sample_info['vt_total'] = response_dict.get("total")
-
-        # Get more information with permalink -------------------------
-        # This is necessary as the VT API does not provide all the needed field values
-        if args.debug:
-            print("[D] Processing permalink {0}".format(response_dict.get("permalink")))
-        info = processPermalink(sample_info["sha256"], args.debug)
-        # Now process the retrieved information
-        # Other info
-        sample_info.update(info)
-        # File Names (special handling)
-        sample_info["filenames"] = ", ".join(info['filenames']).replace(';', '_')
-        sample_info["first_submitted"] = info['firstsubmission']
-
-    return sample_info
-
-
-def processPermalink(sha256, debug=False):
-    """
-    Requests the HTML page for the sample and extracts other useful data
-    that is not included in the public API
-    """
-    headers = {'User-Agent': 'Mozilla/5.0 (compatible; MSIE 10.0; Windows NT 6.1; Trident/6.0)',
-               'Referrer': 'https://www.virustotal.com/',
-               'accept': 'application/json',
-               'accpet-Encoding': 'gzip, deflate, br',
-               'x-app-hostname': 'https://www.virustotal.com/gui/',
-               'x-app-version': '20190717t164815',
-               'Cookie': 'euConsent=1; VT_PREFERRED_LANGUAGE=en'}
-    info = {'filenames': ['-'], 'firstsubmission': '-', 'harmless': False, 'signed': False, 'revoked': False,
-            'expired': False, 'mssoft': False, 'imphash': '-', 'filetype': '-', 'signer': '-',
-            'copyright': '-', 'description': '-', 'comments': 0, 'commenter': ['-'], 'tags': [],
-            'times_submitted': 0, 'reputation': 0, 'filesize': 0}
-
-    try:
-        r_code_details = requests.get(VT_DETAILS % sha256, headers=headers, proxies=PROXY)
-        if r_code_details.status_code != 200:
-             if 'reCAPTCHA' in r_code_details.content.decode("utf-8"):
-                 print("VT_reCAPTCHA issue: Access any report with your browser to unblock your IP, e.g. https://tinyurl.com/vtlup")
-        r_details = json.loads(r_code_details.content.decode("utf-8"))
-
-        #print(json.dumps(r_details, indent=4, sort_keys=True))
-
-        # Get file names
-        info['filenames'] = r_details['data']['attributes']['names']
-        info['filenames'].insert(0, r_details['data']['attributes']['meaningful_name'])
-        # Get file type
-        info['filetype'] = r_details['data']['attributes']['type_description']
-        # Get file size
-        info['filesize'] = convertSize(r_details['data']['attributes']['size'])
-        # Get tags
-        info['tags'] = r_details['data']['attributes']['tags']
-        # First submission
-        info['firstsubmission'] = datetime.utcfromtimestamp(
-            r_details['data']['attributes']['first_submission_date']
-        ).strftime('%Y-%m-%d %H:%M:%S')
-        # Times submitted
-        info['times_submitted'] = r_details['data']['attributes']['times_submitted']
-        # Reputation
-        info['reputation'] = r_details['data']['attributes']['reputation']
-
-        # Exiftool
-        if 'exiftool' in r_details['data']['attributes']:
-            if 'LegalCopyright' in r_details['data']['attributes']['exiftool']:
-                # Get copyright
-                if 'LegalCopyright' in r_details['data']['attributes']['exiftool']:
-                    info['copyright'] = r_details['data']['attributes']['exiftool']['LegalCopyright']
-                # Get description
-                if 'FileDescription' in r_details['data']['attributes']['exiftool']:
-                    info['description'] = r_details['data']['attributes']['exiftool']['FileDescription']
-        # PE Info
-        if 'pe_info' in r_details['data']['attributes']:
-            # Get additional information
-            info['imphash'] = r_details['data']['attributes']['pe_info']['imphash']
-        # PE Signature
-        if 'signature_info' in r_details['data']['attributes']:
-            # Signer
-            if 'signers' in r_details['data']['attributes']['signature_info']:
-                info['signer'] = r_details['data']['attributes']['signature_info']['signers']
-            # Valid
-            if 'verified' in r_details['data']['attributes']['signature_info']:
-                if r_details['data']['attributes']['signature_info']['verified'] == "Signed":
-                    info['signed'] = True
-                if "Revoked" in r_details['data']['attributes']['signature_info']['verified']:
-                    info['revoked'] = True
-                if "Expired" in r_details['data']['attributes']['signature_info']['verified']:
-                    info['expired'] = True
-
-        # Comments
-        r_code_comments = requests.get(VT_COMMENT_API % sha256, headers=headers, proxies=PROXY)
-        r_comments = json.loads(r_code_comments.content.decode("utf-8"))
-        #print(json.dumps(r_comments, indent=4, sort_keys=True))
-        info['comments'] = len(r_comments['data'])
-        if len(r_comments['data']) > 0:
-            info['commenter'] = []
-            for com in r_comments['data']:
-                info['commenter'].append(com['relationships']['author']['data']['id'])
-
-        # Harmless - TODO: legacy features
-        if "Probably harmless!" in r_code_details:
-            info['harmless'] = True
-        # Microsoft Software
-        if "This file belongs to the Microsoft Corporation software catalogue." in r_code_details:
-            info['mssoft'] = True
-    except Exception as e:
-        if args.debug:
-            traceback.print_exc()
-    finally:
-        # Return the info dictionary
-        return info
-
-
-def commentVTSample(resource, comment):
-    """
-    Posts a comment on a certain sample
-    :return:
-    """
-    params = {
-        'apikey': VT_PUBLIC_API_KEY,
-        'resource': resource,
-        'comment': comment
-    }
-    response = requests.post('https://www.virustotal.com/vtapi/v2/comments/put', params=params, proxies=PROXY)
-    response_json = response.json()
-    if response_json['response_code'] != 1:
-        print("[E] Error posting comment: %s" % response_json['verbose_msg'])
-    else:
-        printHighlighted("SUCCESSFULLY COMMENTED")
-
-
 def getMalShareInfo(hash):
     """
     Retrieves information from MalwareShare https://malshare.com
@@ -531,7 +276,7 @@ def getMalShareInfo(hash):
     parameters_query = {"query": hash, "api_key": MAL_SHARE_API_KEY, "action": 'search'}
     parameters_details = {"hash": hash, "api_key": MAL_SHARE_API_KEY, "action": 'details'}
     try:
-        response_query = requests.get(MAL_SHARE_URL, params=parameters_query, timeout=3, proxies=PROXY)
+        response_query = requests.get(MAL_SHARE_URL, params=parameters_query, timeout=3, proxies=connections.PROXY)
         if args.debug:
             print("[D] Querying Malshare: %s" % response_query.request.url)
         #print response_query.content.rstrip('\n')
@@ -540,7 +285,7 @@ def getMalShareInfo(hash):
             info['malshare_available'] = True
             parameters_details['hash'] = response_query.content.decode("utf-8").rstrip('\n')
             #print parameters_details
-            response_details = requests.get(MAL_SHARE_URL, params=parameters_details, proxies=PROXY)
+            response_details = requests.get(MAL_SHARE_URL, params=parameters_details, proxies=connections.PROXY)
             #print response_details.content
         else:
             info['malshare_available'] = False
@@ -639,7 +384,7 @@ def getHybridAnalysisInfo(hash):
         if args.debug:
             print("[D] Querying Hybrid Analysis")
         response = requests.post(HYBRID_ANALYSIS_URL, headers=headers,
-                                timeout=4, proxies=PROXY, data=data)
+                                timeout=4, proxies=connections.PROXY, data=data)
         res_json = response.json()
         # If response has content
         info['hybrid_available'] = len(res_json) > 0
@@ -683,7 +428,7 @@ def getValhalla(sha256):
             "sha256": sha256,
             "apikey": VALHALLA_API_KEY,
         }
-        response = requests.post(VALHALLA_URL, data=data, proxies=PROXY)
+        response = requests.post(VALHALLA_URL, data=data, proxies=connections.PROXY)
         if args.debug:
             print("[D] VALHALLA Response: '%s'" % response.json())
         res = response.json()
@@ -716,7 +461,7 @@ def downloadHybridAnalysisSample(hash):
         # Querying Hybrid Analysis
         if args.debug:
             print("[D] Requesting Downloadsample: %s" % preparedURL)
-        response = requests.get(preparedURL, params={'environmentId':'100'}, headers=headers, proxies=PROXY)
+        response = requests.get(preparedURL, params={'environmentId':'100'}, headers=headers, proxies=connections.PROXY)
 
         # If the response is a json file
         if response.headers["Content-Type"] == "application/json":
@@ -765,7 +510,7 @@ def getTotalHashInfo(sha1):
         # Querying Hybrid Analysis
         if args.debug:
             print("[D] Querying Totalhash: %s" % preparedURL)
-        response = requests.get(preparedURL, proxies=PROXY)
+        response = requests.get(preparedURL, proxies=connections.PROXY)
         # print "Respone: '%s'" % response.content
         if response.content and \
                         '0 of 0 results' not in response.content and \
@@ -797,7 +542,7 @@ def getURLhaus(md5, sha256):
             data = {"sha256_hash": sha256}
         else:
             data = {"md5_hash": md5}
-        response = requests.post(URL_HAUS_URL, data=data, timeout=3, proxies=PROXY)
+        response = requests.post(URL_HAUS_URL, data=data, timeout=3, proxies=connections.PROXY)
         # print("Respone: '%s'" % response.json())
         res = response.json()
         if res['query_status'] == "ok" and res['md5_hash']:
@@ -826,7 +571,7 @@ def getCAPE(md5):
         return info
     try:
         data = {"option": "md5", "argument": md5}
-        response = requests.post(URL_CAPE, data=data, timeout=3, proxies=PROXY)
+        response = requests.post(URL_CAPE, data=data, timeout=3, proxies=connections.PROXY)
         # print("Response: '%s'" % response.json())
         res = response.json()
         if not res['error'] and len(res['data']) > 0:
@@ -849,12 +594,11 @@ def getAnyRun(sha256):
     if sha256 == "-":
         return info
     try:
-        #global PROXY
         
         if args.debug:
             print("[D] Querying Anyrun")
         cfscraper = cfscrape.create_scraper()
-        response = cfscraper.get(URL_ANYRUN % sha256, proxies=PROXY)
+        response = cfscraper.get(URL_ANYRUN % sha256, proxies=connections.PROXY)
        
 
         if args.debug:
@@ -887,7 +631,7 @@ def getVirusBayInfo(hash):
         preparedURL = "%s%s" % (VIRUSBAY_URL, hash)
         if args.debug:
             print("[D] Querying Virusbay: %s" % preparedURL)
-        response = requests.get(preparedURL, proxies=PROXY).json()
+        response = requests.get(preparedURL, proxies=connections.PROXY).json()
         # If response has the correct content
         info['virusbay_available'] = False
         #print(response)
@@ -1041,130 +785,6 @@ def platformChecks(info):
         if args.debug:
             traceback.print_exc()
 
-
-def printResult(info, count, total):
-    """
-    prints the result block
-    :param info: all collected info
-    :param count: counter (number of samples checked)
-    :param total: total number of lines to check
-    :return:
-    """
-    # Rating and Color
-    info["rating"] = "unknown"
-    info["res_color"] = Back.CYAN
-
-    # If VT returned results
-    if "vt_total" in info:
-        info["rating"] = "clean"
-        info["res_color"] = Back.GREEN
-        if info["vt_positives"] > 0:
-            info["rating"] = "suspicious"
-            info["res_color"] = Back.YELLOW
-        if info["vt_positives"] > 10:
-            info["rating"] = "malicious"
-            info["res_color"] = Back.RED
-
-    # Head line
-    printSeparator(count, total, info['res_color'], info["rating"])
-    printHighlighted("HASH: {0} COMMENT: {1}".format(info["hash"], info['comment']))
-
-    # More VT info
-    if "vt_total" in info:
-        # Result
-        info["result"] = "%s / %s" % (info["vt_positives"], info["vt_total"])
-        if info["virus"] != "-":
-            printHighlighted("VIRUS: {0}".format(info["virus"]))
-        printHighlighted("TYPE: {1} SIZE: {2} FILENAMES: {0}".format(removeNonAsciiDrop(info["filenames"]),
-                                                                     info['filetype'],
-                                                                     info['filesize']))
-
-        # Tags to show
-        tags = ""
-        if isinstance(info['tags'], list):
-            tags = " ".join(map(lambda x: x.upper(), info['tags']))
-        # Extra Info
-        printPeInfo(info)
-        printHighlighted("FIRST: {0} LAST: {1} SUBMISSIONS: {5} REPUTATION: {6}\nCOMMENTS: {2} USERS: {3} TAGS: {4}".format(
-            info["first_submitted"],
-            info["last_submitted"],
-            info["comments"],
-            ', '.join(info["commenter"]),
-            tags,
-            info["times_submitted"],
-            info["reputation"]
-        ), tag_color=True)
-
-    else:
-        if args.debug:
-            printHighlighted("VERBOSE_MESSAGE: %s" % info['vt_verbose_msg'])
-
-    # Print the highlighted result line
-    printHighlighted("RESULT: %s" % (info["result"]), hl_color=info["res_color"])
-
-
-def printHighlighted(line, hl_color=Back.WHITE, tag_color=False):
-    """
-    Print a highlighted line
-    """
-    if tag_color:
-        # Tags
-        colorer = re.compile('(HARMLESS|SIGNED|MS_SOFTWARE_CATALOGUE|MSSOFT|SUCCESSFULLY\sCOMMENTED)', re.VERBOSE)
-        line = colorer.sub(Fore.BLACK + Back.GREEN + r'\1' + Style.RESET_ALL + '', line)
-        colorer = re.compile('(REVOKED|EXPLOIT|CVE-[0-9\-]+|OBFUSCATED|RUN\-FILE)', re.VERBOSE)
-        line = colorer.sub(Fore.BLACK + Back.RED + r'\1' + Style.RESET_ALL + '', line)
-        colorer = re.compile('(EXPIRED|VIA\-TOR|OLE\-EMBEDDED|RTF|ATTACHMENT|ASPACK|UPX|AUTO\-OPEN|MACROS)', re.VERBOSE)
-        line = colorer.sub(Fore.BLACK + Back.YELLOW + r'\1' + Style.RESET_ALL + '', line)
-    # Extras
-    colorer = re.compile('(\[!\])', re.VERBOSE)
-    line = colorer.sub(Fore.BLACK + Back.LIGHTMAGENTA_EX + r'\1' + Style.RESET_ALL + '', line)
-    # Add line breaks
-    colorer = re.compile('(ORIGNAME:)', re.VERBOSE)
-    line = colorer.sub(r'\n\1', line)
-    # Standard
-    colorer = re.compile('([A-Z_]{2,}:)\s', re.VERBOSE)
-    line = colorer.sub(Fore.BLACK + hl_color + r'\1' + Style.RESET_ALL + ' ', line)
-    print(line)
-
-
-def printSeparator(count, total, color, rating):
-    """
-    Print a separator line status infos
-    :param count:
-    :param total:
-    :return:
-    """
-    print(Fore.BLACK + color)
-    print(" {0} / {1} > {2}".format(count+1, total, rating.title()).ljust(80) + Style.RESET_ALL)
-
-
-def printKeyLine(line):
-    """
-    Print a given string as a separator line
-    :param line:
-    :return:
-    """
-    print(Fore.BLACK + Back.WHITE)
-    print("{0}".format(line).ljust(80) + Style.RESET_ALL)
-    print("")
-
-
-def printPeInfo(sample_info):
-    """
-    Prints PE information in a clever form
-    :param peInfo:
-    :return:
-    """
-    peInfo = [u'origname', u'description', u'copyright', u'signer']
-    outString = []
-    for k, v in viewitems(sample_info):
-        if k in peInfo:
-            if v is not '-':
-                outString.append("{0}: {1}".format(k.upper(), removeNonAsciiDrop(v)))
-    if " ".join(outString):
-        printHighlighted(" ".join(outString))
-
-
 def saveCache(cache, fileName):
     """
     Saves the cache database as pickle dump to a file
@@ -1189,24 +809,6 @@ def loadCache(fileName):
         # traceback.print_exc()
         return [], False
 
-
-def removeNonAsciiDrop(string):
-    nonascii = "error"
-    # print "CON: ", string
-    try:
-        # Generate a new string without disturbing characters and allow new lines
-        # Python 2 method
-        try:
-            nonascii = "".join(i for i in string if (ord(i) < 127 and ord(i) > 31) or ord(i) == 10 or ord(i) == 13)
-        except Exception as e:
-            # Python 3 fallback
-            return string
-    except Exception as e:
-        traceback.print_exc()
-        pass
-    return nonascii
-
-
 def inCache(hashVal):
     """
     Check if a sample with a certain hash has already been checked and return the info if true
@@ -1219,72 +821,6 @@ def inCache(hashVal):
         if c['hash'] == hashVal or c['md5'] == hashVal or c['sha1'] == hashVal or c['sha256'] == hashVal:
             return c
     return None
-
-
-def knownValue(infos, key, value):
-    """
-    Checks if a value of a given key appears in a previously checked info and
-    returns the info if the value matches
-    :param infos:
-    :param key:
-    :param value:
-    :return: info
-    """
-    for i in infos:
-        if i[key] == value:
-            return i
-    return None
-
-
-def writeCSV(info, resultFile):
-    """
-    Write info line to CSV
-    :param info:
-    :return:
-    """
-    try:
-        with codecs.open(resultFile, 'a', encoding='utf8') as fh_results:
-            # Print every field from the field list to the output file
-            for field_pretty in CSV_FIELD_ORDER:
-                field = CSV_FIELDS[field_pretty]
-                try:
-                    field = info[field]
-                except KeyError as e:
-                    field = "False"
-                try:
-                    field = str(field).replace(r'"', r'\"').replace("\n", " ")
-                except AttributeError as e:
-                    if args.debug:
-                        traceback.print_exc()
-                    pass
-                fh_results.write("%s;" % field)
-            # Append vendor scan results
-            for vendor in VENDORS:
-                if vendor in info['vendor_results']:
-                    fh_results.write("%s;" % info['vendor_results'][vendor])
-                else:
-                    fh_results.write("-;")
-            fh_results.write('\n')
-    except:
-        if args.debug:
-            traceback.print_exc()
-        return False
-    return True
-
-
-def writeCSVHeader(resultFile):
-    """
-    Writes a CSV header line into the results file
-    :param resultFile:
-    :return:
-    """
-    try:
-        with open(resultFile, 'w') as fh_results:
-            fh_results.write("%s;" % ";".join(CSV_FIELD_ORDER))
-            fh_results.write("%s;\n" % ";".join(VENDORS))
-    except Exception as e:
-        print("[E] Cannot write export file {0}".format(resultFile))
-
 
 def getFileData(filePath):
     """
@@ -1301,23 +837,6 @@ def getFileData(filePath):
         traceback.print_exc()
     finally:
         return fileData
-
-
-def convertSize(size_bytes):
-    """
-    Converts number of bytes to a readable form
-    Source: https://stackoverflow.com/questions/5194057/better-way-to-convert-file-sizes-in-python
-    :param size_bytes:
-    :return:
-    """
-    if size_bytes == 0:
-       return "0B"
-    size_name = ("B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB")
-    i = int(math.floor(math.log(size_bytes, 1024)))
-    p = math.pow(1024, i)
-    s = round(size_bytes / p, 2)
-    return "%s %s" % (s, size_name[i])
-
 
 def generateHashes(fileData):
     """
@@ -1420,16 +939,15 @@ if __name__ == '__main__':
     config = configparser.ConfigParser()
     try:
         config.read(args.i)
-        VT_PUBLIC_API_KEY = config['DEFAULT']['VT_PUBLIC_API_KEY']
+        munin_vt.VT_PUBLIC_API_KEY = config['DEFAULT']['VT_PUBLIC_API_KEY']
         MAL_SHARE_API_KEY = config['DEFAULT']['MAL_SHARE_API_KEY']
         PAYLOAD_SEC_API_KEY = config['DEFAULT']['PAYLOAD_SEC_API_KEY']
         VALHALLA_API_KEY = config['DEFAULT']['VALHALLA_API_KEY']
         try:
-            PROXY = config['DEFAULT']['PROXY']
+            connections.setProxy(config['DEFAULT']['PROXY'])
         except KeyError as e:
             print("[E] Your config misses the PROXY field - check the new munin.ini template and add it to your "
                   "config to avoid this error.")
-            PROXY = "-"
 
         # MISP config
         fall_back = False
@@ -1454,21 +972,12 @@ if __name__ == '__main__':
               "changed" % args.i)
 
     # Check API Key
-    if VT_PUBLIC_API_KEY == '' or not re.match(r'[a-fA-F0-9]{64}', VT_PUBLIC_API_KEY):
+    if munin_vt.VT_PUBLIC_API_KEY == '' or not re.match(r'[a-fA-F0-9]{64}', munin_vt.VT_PUBLIC_API_KEY):
         print("[E] No Virustotal API Key set or wrong format")
         print("    Include your API key in a custom config file and use munin.ini as a template\n")
         print("    More info:")
         print("    https://github.com/Neo23x0/munin#get-the-api-keys-used-by-munin\n")
         sys.exit(1)
-
-    # Create valid ProxyDict 
-    if PROXY != "-":
-        proxy_string = PROXY
-        PROXY = {'http': proxy_string, 'https': proxy_string}
-    # No proxy if nothing is set in .ini
-    else:
-        PROXY = {}
-
 
     # Trying to load cache from JSON dump
     cache = []
